pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "savant"
version = "1.0.0"
description = "SAVANT - Semi-automated video annotation toolkit for UAV object detection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "AGPL-3.0"}
authors = [
    {name = "RISE Research Institutes of Sweden"},
]
dependencies = [
    "ultralytics",
    "lap>=0.5.12",
    "opencv-contrib-python>=4.5.0",
    "numpy",
    "pyyaml",
    "pydantic>=2.0",
    "rdflib",
    "dataprov @ git+ssh://git@github.com/fwrise/dataprov.git",
    "PyQt6>=6.5.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "jsonschema>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "jsonschema>=4.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["savant_common", "markit", "trainit", "utils"]

[tool.pytest.ini_options]
testpaths = ["markit/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=markitlib --cov-report=term-missing"

[project.scripts]
markit = "markit.markit:main"
train-yolo-obb = "trainit.train_yolo_obb:main"
split-train-val = "trainit.split_train_val:main"
extract-yolo-training = "trainit.extract_yolo_training:main"
trainit-gui = "trainit.trainit_gui.main:main"
remap-classes-to-ontology = "utils.remap_classes_to_ontology:main"
print-ontology = "utils.print_ontology:main"
